+++ SwiftProject/GameCharacters/Monster.swift
@@ -8,20 +8,13 @@
 import Foundation
 
 class Monster: GameCharacter, CustomStringConvertible {
-    let attackPower: Int
+    let attackPower: Double
     var description: String
     
-    override init(name: String, health: Int, damage: Int) {
+    override init(name: String, health: Double, damage: Double) {
         self.attackPower = damage
         self.description = name.uppercased()
         super.init(name: name, health: health, damage: damage)
     }
     
-//    init(monsterName: String, monsterHealth: Int, attackPower: Int) {
-//        self.monsterName = monsterName
-//        self.monsterHealth = monsterHealth
-//        self.attackPower = attackPower
-//        description = ""
-//        super.init(name: monsterName, health: monsterHealth, damage: attackPower)
-//    }
 }
